Started work on card animation
Added a card-style animation that is pretty close to the vision we had. I may have screwed up some of the coversations, specifically the individual ones as they now all appear as groupchats. But the top coversation "the fellas" works well now. The animation style isn't my favorite and there are "previous" and "next" buttons that can be removed.